Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -7,7 +7,7 @@ int main(){
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
-    float avg = (a+b+c)/3;
+    float avg = (float)(a+b+c)/3;
     printf("Average: %.2f\n", avg);
     return 0;
 }
